Add smart Shards version installation
This uses an exhaustive list of Crystal versions and maps them to the
version of Shards that should be built to accompany it. I've made my
best judgements based on the release notes of Shards, but sometimes it's
hard to tell which versions of Crystal support which versions of Shards.

I also skip support for installing Shards for versions of Crystal
earlier than 0.9.0, as the Shards GitHub releases page does not have
any earlier releases.